Updated preliminary SPI implementation for STM32 variant

+++ src/hw-spi.h
@@ -34,7 +34,7 @@ extern "C" {
 void spi_init(void);
 void spi_deinit(void);
 
-uint16_t spi_transfer(uint16_t data);
+uint8_t spi_transfer(uint8_t data);
 
 #ifdef __cplusplus
 } /* extern "C" */
